mover creation of new anecdote form to a component of its own.
