+++ part6/redux-anecdotes/src/components/AnecdoteForm.js
@@ -1,0 +1,29 @@
+import { useDispatch } from 'react-redux'
+import { createNewAnecdote } from '../reducers/anecdoteReducer'
+
+const AnecdoteForm = () => {
+  const dispatch = useDispatch()
+
+  // add new anecdotes
+
+  const addAnecdotes = (event) => {
+    event.preventDefault()
+    const content = event.target.anecdote.value
+    event.target.anecdote.value = ''
+    dispatch(createNewAnecdote(content))
+  }
+
+  return (
+    <div id='form'>
+      <h2>create new</h2>
+      <form onSubmit={addAnecdotes}>
+        <div>
+          <input name='anecdote' />
+        </div>
+        <button type='submit'>create</button>
+      </form>
+    </div>
+  )
+}
+
+export default AnecdoteForm
